+++ packages/common/src/index.ts
@@ -1,4 +1,4 @@
-// @wellwe/common - 공유 타입/스키마/유틸리티
+// @tryve/common - 공유 타입/스키마/유틸리티
 
 export * from './types';
 export * from './naming-schema';